Feat: Payload v1

+++ Core/Inc/Tasks/sdcard.h
@@ -6,7 +6,7 @@
 
 bool DataStore_SDCardInit(void);
 void DataStore_TelemetrySnapshot(void);
-uint8_t DataStore_WriteToCSV(void);
+int DataStore_WriteToCSV(void);
 void DataStore_WriteToSDCard(int len);
 void send_to_radio(void);
